// src/app/models/deductionReport.ts
export interface deductionReport {
  totalPayCheckDeduction: number;
  dependentsPayCheckDeduction: number;
  employeePayCheckDeduction: number;
  employeeAnnualDeduction: number;
  dependentsAnnualDeduction: number;
  totalAnnualDeduction: number;
  employeePaycheckAfterDeductions: number;
  employeeAnnualPayAfterDeductions: number;
}